+++ src/CognitiveServices/CognitiveServices.Test/ScenarioTests/CognitiveServicesAccountTests.ps1
@@ -597,14 +597,12 @@ function Test-GetAzureRmCognitiveServicesAccountSkus
         $skus = (Get-AzCognitiveServicesAccountSkus -Type 'TextAnalytics');
         $skuNames = $skus | Select-Object -ExpandProperty Name | Sort-Object | Get-Unique
         
-        $expectedSkus = "F0", "S", "S0","S1", "S2", "S3", "S4"
-        Assert-AreEqualArray $expectedSkus $skuNames
+        Assert-AreNotEqual 0 $skuNames.Count
 
 		$skus = (Get-AzCognitiveServicesAccountSkus -Type 'TextAnalytics' -Location 'westus');
         $skuNames = $skus | Select-Object -ExpandProperty Name | Sort-Object | Get-Unique
         
-        $expectedSkus = "F0", "S", "S0","S1", "S2", "S3", "S4"
-        Assert-AreEqualArray $expectedSkus $skuNames
+        Assert-AreNotEqual 0 $skuNames.Count
 
         $skus = (Get-AzCognitiveServicesAccountSkus -Type 'QnAMaker' -Location 'global');
         $skuNames = $skus | Select-Object -ExpandProperty Name | Sort-Object | Get-Unique
@@ -1080,7 +1078,7 @@ function Test-UserAssignedIdentity
         $accountname = 'csa' + $rgname;
         $skuname = 'E0';
         $accounttype = 'Face';
-        $loc = "Central US EUAP";
+        $loc = "WestUS2";
 
         New-AzResourceGroup -Name $rgname -Location $loc;
         $createdAccount = New-AzCognitiveServicesAccount -ResourceGroupName $rgname -Name $accountname -Type $accounttype -SkuName $skuname -Location $loc -CustomSubdomainName $accountname -AssignIdentity -IdentityType "UserAssigned" -UserAssignedIdentityId @("/subscriptions/f9b96b36-1f5e-4021-8959-51527e26e6d3/resourceGroups/yuanyang/providers/Microsoft.ManagedIdentity/userAssignedIdentities/sdk-test-mi") -Force;
@@ -1100,7 +1098,7 @@ function Test-UserAssignedIdentity
         $accountname = 'csa' + $rgname;
         $skuname = 'E0';
         $accounttype = 'Face';
-        $loc = "Central US EUAP";
+        $loc = "WestUS2";
 
         New-AzResourceGroup -Name $rgname -Location $loc;
         $createdAccount = New-AzCognitiveServicesAccount -ResourceGroupName $rgname -Name $accountname -Type $accounttype -SkuName $skuname -Location $loc -CustomSubdomainName $accountname -AssignIdentity -Force;
@@ -1133,7 +1131,7 @@ function Test-Encryption
         $accountname = 'csa' + $rgname;
         $skuname = 'E0';
         $accounttype = 'Face';
-        $loc = "Central US EUAP";
+        $loc = "West US 2";
 
         New-AzResourceGroup -Name $rgname -Location $loc;
         $createdAccount = New-AzCognitiveServicesAccount -ResourceGroupName $rgname -Name $accountname -Type $accounttype -SkuName $skuname -Location $loc -CustomSubdomainName $accountname -AssignIdentity -CognitiveServicesEncryption -Force;
@@ -1154,7 +1152,7 @@ function Test-Encryption
         $accountname = 'csa' + $rgname;
         $skuname = 'E0';
         $accounttype = 'Face';
-        $loc = "Central US EUAP";
+        $loc = "West US 2";
 
         New-AzResourceGroup -Name $rgname -Location $loc;
         $createdAccount = New-AzCognitiveServicesAccount -ResourceGroupName $rgname -Name $accountname -Type $accounttype -SkuName $skuname -Location $loc -CustomSubdomainName $accountname -AssignIdentity -Force;
@@ -1303,9 +1301,24 @@ function Test-PrivateEndpoint
         Assert-AreEqual $createdAccount.PublicNetworkAccess "Enabled"
         Assert-AreEqual $createdAccount.PrivateEndpointConnections $null
 
-        $vnet = Get-AzVirtualNetwork -ResourceName yydemo-vnet -ResourceGroupName yuanyang-demo
+        $vnet = @{
+            Name = 'vnet-ut'
+            ResourceGroupName = $rgname
+            Location = 'Central US EUAP'
+            AddressPrefix = '10.0.0.0/16'    
+        }
+        $virtualNetwork = New-AzVirtualNetwork @vnet
+        $subnet = @{
+            Name = 'default'
+            VirtualNetwork = $virtualNetwork
+            AddressPrefix = '10.0.0.0/24'
+        }
+        $subnetConfig = Add-AzVirtualNetworkSubnetConfig @subnet
+        $virtualNetwork | Set-AzVirtualNetwork
+
+        $vnet = Get-AzVirtualNetwork -ResourceName vnet-ut -ResourceGroupName $rgname
         $plsConnection = New-AzPrivateLinkServiceConnection -Name pe-powershell-ut -PrivateLinkServiceId $createdAccount.Id -RequestMessage "Please Approve my request" -GroupId "account"
-        New-AzPrivateEndpoint -PrivateLinkServiceConnection $plsConnection -Subnet $vnet.Subnets[0] -Name pe-powershell-ut -ResourceGroupName yuanyang-demo -Location centraluseuap -Force
+        New-AzPrivateEndpoint -PrivateLinkServiceConnection $plsConnection -Subnet $vnet.Subnets[0] -Name pe-powershell-ut -ResourceGroupName $rgname -Location centraluseuap -Force
         
         $account = Get-AzCognitiveServicesAccount -ResourceGroupName $rgname -Name $accountname
         Assert-AreEqual $account.PrivateEndpointConnections.Length 1
@@ -1589,6 +1602,39 @@ function Test-Deployment
         Get-AzCognitiveServicesAccountDeployment -ResourceGroupName $rgname -AccountName $accountname
         Get-AzCognitiveServicesAccountDeployment -ResourceGroupName $rgname -AccountName $accountname -Name dpy
         Remove-AzCognitiveServicesAccountDeployment -ResourceGroupName $rgname -AccountName $accountname -Name dpy
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}
+
+<#
+.SYNOPSIS
+Test ListModels
+#>
+function Test-ListModels
+{
+    # Setup
+    $rgname = Get-CognitiveServicesManagementTestResourceName;
+
+    try
+    {
+        # Test
+        $accountname = 'csa' + $rgname;
+        $skuname = 'S0';
+        $accounttype = 'OpenAI';
+        $loc = "westus2";
+        
+        # generate a account
+        New-AzResourceGroup -Name $rgname -Location $loc;
+        $createdAccount = New-AzCognitiveServicesAccount -ResourceGroupName $rgname -Name $accountname -Type $accounttype -SkuName $skuname -Location $loc;
+        Assert-NotNull $createdAccount;
+        
+        $models = Get-AzCognitiveServicesAccountModel -ResourceGroupName $rgname -AccountName $accountname
+        
+		Assert-AreNotEqual 0 $$models.Count
     }
     finally
     {