+++ src/Kusto/test/Update-AzKustoScript.Tests.ps1
@@ -15,8 +15,21 @@ Describe 'Update-AzKustoScript' {
         }
         . ($mockingPath | Select-Object -First 1).FullName
     }
-    It 'UpdateExpanded' -skip {
-        { throw [System.NotImplementedException] } | Should -Not -Throw
+    It 'UpdateExpanded' {
+        $continueOnErrors = $false
+        $clusterName = $env.clusterName
+        $databaseName = "testdatabase" + $env.rstr4
+        $databaseFullName = $clusterName + "/" + $databaseName
+        $scriptContent = ".create table table3 (Level:string, Timestamp:datetime, UserId:string, TraceId:string, Message:string, ProcessId:int32)"
+
+        $databaseCreated = New-AzKustoDatabase -ResourceGroupName $env.resourceGroupName -ClusterName $clusterName -Name $databaseName -Kind ReadWrite -Location $env.location
+        
+        $Script = New-AzKustoScript -ClusterName $clusterName -DatabaseName $databaseName -Name $env.scriptName -ResourceGroupName $env.resourceGroupName -SubscriptionId $env.SubscriptionId -ContinueOnError -ForceUpdateTag $env.forceUpdateTag -ScriptContent $scriptContent
+        $Script = Update-AzKustoScript -ClusterName $clusterName -DatabaseName $databaseName -Name $env.scriptName -ResourceGroupName $env.resourceGroupName -SubscriptionId $env.SubscriptionId -ContinueOnError -ForceUpdateTag $env.forceUpdateTag2 -ScriptContent $scriptContent
+        
+        Validate_Inline_Script $Script $env.forceUpdateTag2 $continueOnErrors $clusterName $databaseName $env.scriptName
+
+        { Remove-AzKustoDatabase -ResourceGroupName $env.resourceGroupName -ClusterName $env.clusterName -Name $name } | Should -Not -Throw
     }
 
     It 'Update' -skip {