+++ src/ApplicationInsights/ApplicationInsights/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.3.1
+* Added breaking change warnings for upcoming Az.ApplicationInsights 2.0.0
+
 ## Version 1.3.0
 * Added WebTest function. Below is the new cmdlet
     * `Get-AzApplicationInsightsWebTest`