+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,20 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported DaysAfterLastTierChangeGreaterThan in Management Policy
+    -  `Add-AzStorageAccountManagementPolicyAction`
+* Fixed the issue that upload blob might fail on Linux [#17743]
+    -  `Set-AzStorageBlobContent`
+* Supported AllowPermanentDelete when enable blob soft delete
+    - `Enable-AzStorageBlobDeleteRetentionPolicy`
+* Added breaking change warning message for upcoming cmdlet breaking change
+    - `Get-AzStorageFile`
+
+## Version 4.4.1
+* Fixed get blob by tag failure on Powershell 7.2.2
+    -  `Get-AzStorageBlobByTag`
+
+## Version 4.4.0
 * Updated examples in reference documentation for `Close-AzStorageFileHandle`
 * Supported create storage context with customized blob, queue, file, table service endpoint
     - `New-AzStorageContext`